functions/firestore/requirements.txt
google-cloud-firestore==2.3.4
uuid==1.30